# backend/internal/handlers/http/docs/swagger.yaml
definitions:
  EmptyResponseDTO:
    properties:
      code:
        example: 200
        type: integer
      message:
        example: Operação realizada com sucesso
        type: string
      status:
        example: success
        type: string
    type: object
  ResponseDTO-UserToken:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/UserToken'
      message:
        example: Operação realizada com sucesso
        type: string
      status:
        example: success
        type: string
    type: object
  UserLoginChangeDTO:
    properties:
      email:
        type: string
      new_password:
        minLength: 6
        type: string
      session:
        type: string
    required:
    - email
    - new_password
    - session
    type: object
  UserLoginDTO:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  UserToken:
    properties:
      AccessToken:
        type: string
      AuthChallenge:
        type: string
      AuthSession:
        type: string
      RefreshToken:
        type: string
    type: object
info:
  contact: {}
  description: Auth3 Rest API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Auth3 Rest API
  version: "1.0"
paths:
  /auth3/change:
    post:
      description: Change password
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UserLoginChangeDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ResponseDTO-UserToken'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Change password
      tags:
      - Auth
  /auth3/login:
    post:
      description: User login
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UserLoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ResponseDTO-UserToken'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: User login
      tags:
      - Auth
  /auth3/status:
    get:
      description: healthcheck
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/EmptyResponseDTO'
      summary: healthcheck
      tags:
      - Auth
swagger: "2.0"